adds database configuration
