+++ database/models/index.cjs
@@ -1,11 +1,21 @@
-'use strict';
+// import fs from 'fs';
+// import path from 'path';
+// import Sequelize from 'sequelize';
+// import { fileURLToPath } from 'url'
+// import configFile from '../config.cjs';
 
+
+// const __filename = fileURLToPath(import.meta.url);
+// const __dirname = path.dirname(__filename);
+// const model_name = 
 const fs = require('fs');
 const path = require('path');
 const Sequelize = require('sequelize');
+const configFile = require('../config.cjs');
+
 const basename = path.basename(__filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require(__dirname + '/../config.js')[env];
+const config = configFile[env];
 const db = {};
 
 let sequelize;
@@ -18,7 +28,7 @@ if (config.use_env_variable) {
 fs
   .readdirSync(__dirname)
   .filter(file => {
-    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-3) === '.js');
+    return (file.indexOf('.') !== 0) && (file !== basename) && (file.slice(-4) === '.cjs');
   })
   .forEach(file => {
     const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
@@ -35,3 +45,4 @@ db.sequelize = sequelize;
 db.Sequelize = Sequelize;
 
 module.exports = db;
+// export default db;
